requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
assemblyai
requests
numpy
asyncio
websocket-client
langchain
langchain_core
langchain_google_genai
google-cloud
google-cloud-dialogflow-cx
google-cloud-speech
pyaudio
pydub
ffmpeg
python-multipart
google-cloud-texttospeech
python-docx
pypdf2
motor